Query module: Separate BlockProcessorApi and MetadataProcessorApi
