+++ packages/core/src/query/interfaces/MetadataProcessorApi.ts
@@ -1,0 +1,9 @@
+import { BlockProcessorApi } from './BlockProcessorApi'
+
+export type MetaTxStatus =
+  | { isSuccess: true }
+  | { isSuccess: false; error: string }
+
+export interface MetadataProcessorApi extends BlockProcessorApi {
+  metaTxStatus(txHash: `0x${string}`): Promise<MetaTxStatus | null>
+}
