+++ packages/core/src/query/interfaces/index.ts
@@ -1,1 +1,2 @@
 export * from './BlockProcessorApi'
+export * from './MetadataProcessorApi'
