+++ packages/core/src/query/storageSquid/consts.ts
@@ -1,0 +1,4 @@
+import { ENTITY_INFO } from './__generated__/entityInfo'
+import { AnyEntity } from './__generated__/QueryApi'
+
+export const ALL_ENTITIES = Object.keys(ENTITY_INFO) as AnyEntity[]
